Fix activity report by having the user pick months
Just jumping back a number of months works fine for the activity _page_,
but not for the activity _report_, since the report actually needs to
deal with months. Instead of trying to be clever with auto-detecting the
months that the user wants, now I just have them pick from the months of
activity that exist in the database with a simple SQL query and multiple
select control.

Given that the user might have mis-selected the months or see data for
a different set of months, I also added a text link and corresponding
endpoint to delete the report from the cache so that the user doesn't
have to wait 6 hours for it to expire.